# Conflicts:
#	app/build.gradle.kts
#	app/src/main/java/com/example/assu_fe_app/data/DevBearerInterceptor.kt
#	app/src/main/java/com/example/assu_fe_app/data/repository/deviceToken/DeviceTokenRepository.kt
#	app/src/main/java/com/example/assu_fe_app/data/repositoryImpl/deviceToken/DeviceTokenRepositoryImpl.kt
#	app/src/main/java/com/example/assu_fe_app/data/service/deviceToken/DeviceTokenService.kt
#	app/src/main/java/com/example/assu_fe_app/di/RepositoryModule.kt
#	app/src/main/java/com/example/assu_fe_app/domain/usecase/deviceToken/UnregisterDeviceTokenUseCase.kt
#	app/src/main/java/com/example/assu_fe_app/fcm/MyFirebaseMessagingService.kt
#	app/src/main/java/com/example/assu_fe_app/presentation/admin/mypage/AdminMypageFragment.kt
#	app/src/main/java/com/example/assu_fe_app/presentation/common/login/LoginActivity.kt
#	app/src/main/java/com/example/assu_fe_app/presentation/partner/mypage/PartnerMypageFragment.kt
#	app/src/main/java/com/example/assu_fe_app/presentation/user/mypage/UserMypageFragment.kt
#	app/src/main/java/com/example/assu_fe_app/ui/deviceToken/DeviceTokenViewModel.kt

+++ app/src/main/java/com/example/assu_fe_app/data/manager/TokenManager.kt
@@ -95,8 +95,8 @@ class TokenManager @Inject constructor(
         return prefs.getLong(KEY_USER_ID, -1L)
     }
     
-    fun saveDeviceTokenId(tokenId: Int) {
-        prefs.edit().putInt(KEY_DEVICE_TOKEN_ID, tokenId).apply()
+    fun saveDeviceTokenId(tokenId: Long) {
+        prefs.edit().putLong(KEY_DEVICE_TOKEN_ID, tokenId).apply()
     }
     
     fun getDeviceTokenId(): Int? {
